+++ src/constants/feeds.ts
@@ -1,4 +1,5 @@
 import publishNpm from '../assets/images/blogs/how-to-make-your-own-npm-package.webp';
+import jsArrayMethods from '../assets/images/blogs/js-array-methods.webp';
 
 const FEEDS: SingleFeed[] = [
 	{
@@ -10,6 +11,13 @@ const FEEDS: SingleFeed[] = [
 		url: 'https://dev.to/mohammedtaysser/how-to-make-your-own-npm-package-4dm2',
 		img: publishNpm,
 		publishAt: new Date('2023-02-21T18:10:24Z'),
+	},
+	{
+		title: 'JavaScript Array Methods: A Guide with Examples ',
+		info: `Arrays are one of the most commonly used data structures in JavaScript. It provides a lot of built-in methods to create, modify, search, and even transform arrays result.`,
+		url: 'https://dev.to/mohammedtaysser/javascript-array-methods-a-guide-with-examples-3a16',
+		img: jsArrayMethods,
+		publishAt: new Date('2023-09-06T19:10:24Z'),
 	},
 ];